+++ packages/create-sitecore-jss/src/common/processes/transform.ts
@@ -196,10 +196,8 @@ export const diffAndWriteFiles = async ({
 export const populateEjsData = (answers: BaseArgs, destination?: string) => {
   // pass in helper to answers object
 
-  // Don't expose canary build number in the generated app
-  const jssVersion = version.includes('canary')
-    ? version.replace(/(-canary\.\d+)$/, '-canary')
-    : version;
+  // Use exact version for jss dependencies in beta and canary versions
+  const jssVersion: string = version.match(/(\-[a-zA-Z]+\.\d+)$/) ? version : `~${version}`;
 
   const ejsData: Data = {
     ...answers,