+++ packages/sitecore-jss-nextjs/src/components/Link.tsx
@@ -1,5 +1,6 @@
 import React, { forwardRef, JSX } from 'react';
 import NextLink from 'next/link';
+import { LinkProps as NextLinkProps } from 'next/link';
 import {
   Link as ReactLink,
   LinkFieldValue,
@@ -13,6 +14,11 @@ export type LinkProps = ReactLinkProps & {
    * @default /^\//g
    */
   internalLinkMatcher?: RegExp;
+
+  /**
+   * Next.js Link prefetch.
+   */
+  prefetch?: NextLinkProps['prefetch'];
 };
 
 /**
@@ -58,6 +64,7 @@ export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
 
       // determine if a link is a route or not. File extensions are not routes and should not be pre-fetched.
       if (isMatching && !isFileUrl) {
+        delete htmlLinkProps.emptyFieldEditingComponent;
         return (
           <NextLink
             href={{ pathname: href, query: querystring, hash: anchor }}
@@ -66,6 +73,7 @@ export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
             title={value.title}
             target={value.target}
             className={value.class}
+            prefetch={props.prefetch}
             {...htmlLinkProps}
             ref={ref}
           >
@@ -76,9 +84,10 @@ export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
       }
     }
 
-    // prevent passing internalLinkMatcher as it is an invalid DOM element prop
+    // prevent passing internalLinkMatcher or prefetch as it is an invalid DOM element prop
     const reactLinkProps = { ...props };
     delete reactLinkProps.internalLinkMatcher;
+    delete reactLinkProps.prefetch;
 
     return <ReactLink {...reactLinkProps} ref={ref} />;
   }