+++ packages/sitecore-jss/src/layout/content-styles.ts
@@ -60,7 +60,7 @@ export const traversePlaceholder = (
 };
 
 export const traverseField = (field: Field | Item | Item[] | undefined, config: Config) => {
-  if (!field || config.loadStyles) return;
+  if (!field || typeof field !== 'object' || config.loadStyles) return;
 
   if ('editable' in field && field.editable) {
     config.loadStyles = CLASS_REGEXP.test(field.editable);